added event package. Moved CRUD functions to event.go. GetEvents route grabs var events from events.go
