+++ pkg/adding/service.go
@@ -1,0 +1,7 @@
+package adding
+
+import "fmt"
+
+func addEvent() {
+	fmt.Println("adding event...")
+}
